OLD_CODE_2/backend_OLD_Django/requirements.dev.txt
pytest
pytest-django
black
mypy
types-redis
celery-types
isort
django-stubs[compatible-mypy]